+++ packages/react-compass/src/dropdown/dropdown.flag.tsx
@@ -21,9 +21,11 @@ interface P<T = object> extends ComboBoxStateOptions<T>, StyledComponentProps {
   isLoading?: boolean
   icon?: React.ReactNode
   isErrored?: boolean
+  isRequired?: boolean
   errorMessage?: string
   headerTitle?: string
-  keyType?: 'alpha-2' | 'alpha-3' | 'name' | 'country-code' | 'phone-code'
+  keyType?: 'alpha-2' | 'alpha-3' | 'name' | 'country-code'
+  onPhoneChange?: (p: string) => void
   headerOnClick?: (e: unknown) => void
 }
 
@@ -49,7 +51,11 @@ const PreDropdown = React.forwardRef<HTMLDivElement, P>((props, ref) => {
     css = {},
     icon = <Icon />,
     isErrored,
+    isRequired,
+    isDisabled,
     errorMessage,
+    placeholder,
+    onPhoneChange,
     // AriaDropdownProps
   } = props
   const variantProps = {} as DropdownVariantProps
@@ -91,13 +97,28 @@ const PreDropdown = React.forwardRef<HTMLDivElement, P>((props, ref) => {
     [inputProps.value],
   )
 
+  React.useEffect(() => {
+    if (currentState?.['phone-code']) {
+      onPhoneChange?.(currentState['phone-code'])
+    }
+  }, [currentState])
+
   return (
     <StyledDropdownWrapper css={css} ref={dropdownRef} {...variantProps}>
-      {props.label && <label {...labelProps}>{props.label}</label>}
-      <StyledFlag isEmpty={!inputProps.value} isErrored={!!isErrored}>
+      {props.label && (
+        <label {...labelProps}>
+          {isRequired && <span>*</span>}
+          {props.label}
+        </label>
+      )}
+      <StyledFlag
+        isEmpty={!inputProps.value && !placeholder}
+        isErrored={!!isErrored}
+        isDisabled={!!isDisabled}
+      >
         {currentState && (
           <StyledFlagIcon>
-            <Flag iso={currentState['alpha-3']} />
+            <Flag iso={currentState['alpha-2']} />
           </StyledFlagIcon>
         )}
         <input {...inputProps} ref={inputRef} />