+++ packages/react-compass/src/dropdown/dropdown.stories.tsx
@@ -8,7 +8,11 @@ export const Flag: React.FC = () => {
   return (
     <Column>
       <h3>Flag</h3>
-      <Dropdown.Flag label='List of country' placeholder='Choose a country' keyType='phone-code' />
+      <Dropdown.Flag
+        label='List of country'
+        placeholder='Choose a country'
+        onPhoneChange={(e) => console.log(e)}
+      />
     </Column>
   )
 }
@@ -21,6 +25,7 @@ export const Select: React.FC = () => {
       <Dropdown.Select
         label='Favorite Animal'
         placeholder='Choose an animal'
+        isDisabled
         selectedKey={value}
         onSelectionChange={(k: Key) => setValue(k)}
       >