+++ packages/react-compass/src/dropdown/flags/nl.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react'
+const NLFlag: React.FC = () => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    id='flag-icons-nl'
+    viewBox='0 0 640 480'
+    width='1em'
+    height='1em'
+  >
+    <path fill='#21468b' d='M0 0h640v480H0z' />
+    <path fill='#fff' d='M0 0h640v320H0z' />
+    <path fill='#ae1c28' d='M0 0h640v160H0z' />
+  </svg>
+)
+export default NLFlag