+++ packages/react-compass/src/textarea/textarea.tsx
@@ -15,9 +15,13 @@ import {
 
 interface Props extends AriaTextFieldProps, StyledComponentProps {
   label?: string
-  disabled?: boolean
-  errored?: boolean
+  isDisabled?: boolean
+  isRequired?: boolean
+  isErrored?: boolean
+  errorMessage?: string
   wordCount?: boolean
+  onChange?: (value: string) => void
+  onChangeEvent?: (event: React.ChangeEvent<HTMLInputElement>) => void
 }
 
 export type TextareaProps = Props &
@@ -31,15 +35,19 @@ const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
       css = {},
       // ComponentProps
       label,
-      disabled,
-      errored,
+      isDisabled,
+      isErrored,
+      isRequired,
+      errorMessage,
       wordCount,
       maxLength,
+      onChange,
+      onChangeEvent,
       // AriaTextFieldProps
       ...ariaSafeProps
     } = props
 
-    const variantProps = {errored} as TextareaVariantProps
+    const variantProps = {isErrored} as TextareaVariantProps
     const ariaProps = {
       label,
       maxLength,
@@ -58,13 +66,30 @@ const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
       keyof Props
     >
 
+    const handleOnChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+      //inputProps.onChange?.(event)
+      onChange?.(event.target.value)
+      onChangeEvent?.(event)
+    }
+
     return (
       <StyledTextareaWrapper css={css} {...htmlProps}>
-        <StyledTextFieldLabel {...labelProps} disabled={!!disabled}>
-          {label}
-        </StyledTextFieldLabel>
-        <StyledTextarea ref={textareaRef} {...inputProps} {...variantProps} />
-        {wordCount ? (
+        {label && (
+          <StyledTextFieldLabel {...labelProps} isDisabled={!!isDisabled}>
+            {isRequired && <span>*</span>} {label}
+          </StyledTextFieldLabel>
+        )}
+        <StyledTextarea
+          {...inputProps}
+          {...variantProps}
+          ref={textareaRef}
+          isErrored={!!isErrored}
+          onChange={(e) => {
+            console.log(e)
+            handleOnChange(e as unknown as React.ChangeEvent<HTMLInputElement>)
+          }}
+        />
+        {wordCount && (
           <StyledTextFieldHelperText
             className='word-count'
             {...descriptionProps}
@@ -72,7 +97,12 @@ const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
             {inputProps.value?.toString().length ?? '0'}
             {maxLength ? `/${maxLength}` : null}
           </StyledTextFieldHelperText>
-        ) : null}
+        )}
+        {isErrored && errorMessage && (
+          <StyledTextFieldHelperText error>
+            {errorMessage}
+          </StyledTextFieldHelperText>
+        )}
       </StyledTextareaWrapper>
     )
   },